LayoutTests/http/wpt/service-workers/fetch-service-worker-preload-download-through-direct-preload.https.html

<!doctype html>
<html>
<head>
<script src="/common/utils.js"></script>
<script src="resources/routines.js"></script>
</head>
<body>
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.setShouldLogDownloadCallbacks(true);
  testRunner.setShouldLogDownloadSize(true);
  testRunner.setShouldLogDownloadExpectedSize(true);
  testRunner.waitUntilDownloadFinished();
  testRunner.setShouldDownloadUndisplayableMIMETypes(true);
}

const url = "/WebKit/service-workers/resources/fetch-service-worker-preload-script.py?download=yes&no-fetch-event-handling=yes&token=" + token();

async function doTest() {
    await setupActivatedButNotRunningServiceWorker("/WebKit/service-workers/fetch-service-worker-preload-worker.js", url);

    await fetch(url + "&value=use-preload", { method: 'POST' });

    // Load iframe, with not running worker, so preload will start and will be the one to be downloaded, directly through preload since service worker did not handle the event.
    // Download size should be size of "use-preload".
    withIframe(url);
}
doTest();
</script>
</body>
</html>
