LayoutTests/http/wpt/service-workers/fetch-service-worker-preload-download-through-direct-preload.https-expected.txt

Download started.
Downloading URL with suggested filename "fetch-service-worker-preload-script.py.vcf"
Download size: 11.
Download expected size: 11.
Download completed.
